# openapi/ge-v1.yaml
openapi: 3.0.3
info:
  title: skstudio Project Governance (ge/org) API
  description: |
    Normative OpenAPI for P1 project governance domain (v1.1.3 · P1 spec frozen).
    Full spec: docs/目标与执行及其组织支撑(实现与测试).md §4.9
    Business rules: docs/目标与执行及其组织支撑(需求).md v2.7.2
    M20: system lifecycle tasks (启动项目/结项复盘/确认结项) · is_system on Task/GateItem
    M20b: 结项复盘 auto-seeded PM sign route (确认结项 prerequisite) · system_sign_route_immutable
    M21: permission simplification · PM=reviewer governance · not_project_governor
    v1.10.0: gate.includes auto-sync · POST/DELETE includes → 410 · Canvas produce/prerequisite only
    v1.10.2: end closure sign route auto-seed · protected prerequisite edge
    Out of scope: skstudio BFF /api/v1/org|ge/* and POST /internal/ge/ws-fanout (see impl doc §4.0 · §5.4)
  version: 1.10.2
servers:
  - url: /api/v1
    description: goal_execution service (JWT participant + service governance via skstudio BFF)
paths:
...
  /ge/programs/{program_id}:
    get:
      operationId: getGeProgram
      summary: Program detail with projects[] (JWT participant filter; reviewer via BFF)
      tags: [ge-programs]
      parameters:
        - $ref: '#/components/parameters/ProgramId'
      responses:
        '200':
          description: Program detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /ge/projects:
    get:
      operationId: listGeProjects
      summary: List projects (JWT participant ∪ draft creator; reviewer via skstudio BFF)
      tags: [ge-projects]
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummary'
    post:
      operationId: createGeProject
      summary: Atomically create project graph (draft)
      tags: [ge-projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /ge/projects/{project_id}:
    get:
      operationId: getGeProject
      summary: Project detail
      tags: [ge-projects]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project detail
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchGeProject
      summary: Update name or pm_user_id (M19 name syncs project Note title via BFF)
      tags: [ge-projects]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchProjectRequest'
      responses:
        '200':
          description: Updated
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteGeProject
      summary: Soft-delete empty project (PM or reviewer governance · project_is_empty)
      tags: [ge-projects]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Soft deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /ge/projects/{project_id}/project-note:
    patch:
      operationId: bindGeProjectNote
      summary: Bind project_note_id for legacy backfill (service token only · GE-T50)
      tags: [ge-projects]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_note_id]
              properties:
                project_note_id:
                  type: string
                  format: uuid
                  description: KB note stable id; only when current value is null
      responses:
        '200':
          description: Project with bound note
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: project_note_already_bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /ge/projects/{project_id}/graph:
    get:
      operationId: getGeProjectGraph
      summary: Project execution graph for Canvas
      tags: [ge-projects]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGraph'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /ge/projects/{project_id}/phases/{phase_id}/tasks:
    post:
      operationId: addGePhaseTask
      summary: Add task to phase (pm; assignee_user_id required)
      description: |
        Creates a blocked task in the given phase. assignee_user_id is required;
        the server does not fall back to pm_user_id when omitted or empty.
      tags: [ge-tasks]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/PhaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPhaseTaskRequest'
      responses:
        '200':
          description: Updated graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGraph'
        '400':
          description: invalid_assignee or invalid_request
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /ge/projects/{project_id}/phases/{phase_id}/gate-items:
    post:
      operationId: addGePhaseGateItem
      summary: Add gate item to phase (pm)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/PhaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPhaseGateItemRequest'
      responses:
        '200':
          description: Updated graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGraph'
        '400':
          description: invalid_request or schedule validation
        '403':
          $ref: '#/components/responses/Forbidden'
  /ge/phases/{phase_id}:
    patch:
      operationId: patchGePhase
      summary: Update business phase name and planned window (pm)
      tags: [ge-phases]
      parameters:
        - $ref: '#/components/parameters/PhaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                planned_start:
                  type: string
                  format: date
                  nullable: true
                planned_end:
                  type: string
                  format: date
                  nullable: true
      responses:
        '200':
          description: Updated graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGraph'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteGePhase
      summary: Delete empty non-active business phase (pm)
      tags: [ge-phases]
      parameters:
        - $ref: '#/components/parameters/PhaseId'
      responses:
        '200':
          description: Updated graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGraph'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /ge/tasks/{task_id}:
    get:
      operationId: getGeTaskContext
      summary: Task context for deep links (M12)
      tags: [ge-tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task with project_id and phase_id
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchGeTask
      summary: Update assignee_user_id, title, or phase_id (pm)
      tags: [ge-tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assignee_user_id:
                  type: string
                title:
                  type: string
                phase_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Updated task
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteGeTask
      summary: Delete blocked task with no links (pm)
      tags: [ge-tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Updated graph
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /ge/tasks/{task_id}/start:
    post:
      operationId: startGeTask
      summary: ready to running (assignee)
      tags: [ge-tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Write operation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperationResponse'
  /ge/tasks/{task_id}/done:
    post:
      operationId: doneGeTask
      summary: running to done (assignee)
      tags: [ge-tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Write operation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperationResponse'
  /ge/gate-items/{gate_item_id}:
    get:
      operationId: getGeGateItemContext
      summary: Gate item context for deep links (M12)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/GateItemId'
      responses:
        '200':
          description: Gate item with project_id, phase_id, gate_id
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchGeGateItem
      summary: Update gate item name, phase_id, planned_due, form, or definition fields (pm; draft only for form)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/GateItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phase_id:
                  type: string
                  format: uuid
                planned_due:
                  type: string
                  format: date
                form:
                  type: string
                  enum: [material, metric, status]
                target_value:
                  oneOf:
                    - type: number
                    - type: string
                    - type: boolean
                operator:
                  type: string
                  enum: ['>=', '<=', '==', '>', '<']
                target_state:
                  type: string
      responses:
        '200':
          description: Updated graph
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteGeGateItem
      summary: Delete draft gate item with no links (pm)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/GateItemId'
      responses:
        '200':
          description: Updated graph
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /ge/gate-items/{gate_item_id}/submit:
    post:
      operationId: submitGeGateItem
      summary: Submit gate item (produce assignee)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/GateItemId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Write operation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperationResponse'
  /ge/gate-items/{gate_item_id}/sign:
    post:
      operationId: signGeGateItem
      summary: Sign gate item (eligible signer)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/GateItemId'
      responses:
        '200':
          description: Write operation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperationResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /ge/gate-items/{gate_item_id}/reject:
    post:
      operationId: rejectGeGateItem
      summary: Reject gate item (eligible signer)
      tags: [ge-gate-items]
      parameters:
        - $ref: '#/components/parameters/GateItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reject_reason]
              properties:
                reject_reason:
                  type: string
                  minLength: 10
                  maxLength: 500
      responses:
        '200':
          description: Write operation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperationResponse'
...
  /ge/me/execution-notifications:
    get:
      operationId: listGeExecutionNotifications
      summary: Phase progress notifications
      tags: [ge-notifications]
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Notification list
...
components:
  parameters:
...
    ProgramId:
      name: program_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Not found or soft-deleted (detail project_not_found)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...
  schemas:
...
    ProgramDetail:
      allOf:
        - $ref: '#/components/schemas/ProgramMetadata'
        - type: object
          required: [projects]
          properties:
            projects:
              type: array
              description: JWT filtered to participant ∪ draft creator
              items:
                $ref: '#/components/schemas/ProjectSummary'
...